requirements.txt
selenium >= 4.3.0
robotframework >= 4.1.3
robotframework-pythonlibcore >= 3.0.0
robotframework-seleniumlibrary>=5.1.3
robotframework-requests>=0.9.2
robotframework-jsonlibrary>=0.4.1
robotframework-robocop>=3.1.1
robotframework-debuglibrary>=2.3.0
robotframework-excellib>=2.0.1
robotframework-imaplibrary2>=0.4.6
robotframework-appiumlibrary>=2.0.0
robotframework-pdf2textlibrary>=1.0.1
robotframework-browsermobproxylibrary>=0.1.3
pyyaml>=5.4.1
pyzbar>=0.1.9
Pillow>=10.0.0
xlrd